pyproject.toml
[tool.poetry]
name = "starnavi-test"
version = "0.1.0"
description = ""
authors = ["Serhii Tolmachov <cvaizencva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.21.4"
django-ninja-jwt = "^5.3.4"
python-dotenv = "^1.0.1"
factory-boy = "^3.3.1"
ninja-schema = "^0.13.6"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.3"
ruff = "^0.7.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
fix = true


[tool.coverage.run]
branch = true
omit = [
    "*/__init__.py",
    "*/tests.py",
    "*/migrations/*",
    "*/settings/*",
    "*/wsgi.py",
    "*/asgi.py",
    "manage.py",
    "*_base.py",
    "test_*.py",
]
source = ["."]


[tool.coverage.report]
show_missing = true
